Rename hidden input component
